refactor some stuff
